Fixed the program not finding audio files. Added task to TODO list.
I finally fixed the problem that caused the program to not be able to
use the audio files. I had to place the files inside of the src
directory, which I didn't want, but it works which is all I need.

When I tried to actually use the audio, it said that it was in an
invalid format. This might just be my testing environment (Linux with
messed up Java), so I'll test it tomorrow with a Windows PC.