Added debugging mode and related functions.
All debugging mode does is allow the program to write to the console.
Specifically it allows the program to write the value of each click to
the console. Since the average user won't run the program from the
console, I thought it would be wasteful to always print stuff to the
console. Now, if you want to see the values of the clicks, run the
program with the "debug" parameter.

I added the functions needed to facilitate the use of parameters in the
program and made it easy to add more parameters, just add another if
statement!